refactor(reporting): moves "Reporting.ts" to "Reporting/promptUserWith.ts"
